Atualizando/Melhorando Funcionalidades da Interface

+++ v1/DB/DataBase.py
@@ -15,32 +15,47 @@ class DATABASE():
         self.cursor = self.DB.cursor()
 
     def visualizarProdutos(self, TAG=None):
-        if TAG is not None:
-            try:
-                TAG=int(TAG)
-                TAG=str(TAG)
-                query = f"SELECT * FROM {self.INFOS["table"]} WHERE ID = %s"
-                self.cursor.execute(query, (TAG, ))
-                produto = self.cursor.fetchall()
-                print(produto)
-                return produto
-            
-            except:
-                query = f"SELECT * FROM {self.INFOS["table"]} WHERE NOME LIKE %s"
-                self.cursor.execute(query, (TAG, ))
-                produto = self.cursor.fetchall()
-                print(produto)
-                return produto
+        try:
+            if TAG is not None:
+                try:
+                    TAG = int(TAG)
+                    TAG = str(TAG)
+                    query = f"SELECT * FROM {self.INFOS['table']} WHERE ID = %s"
+                    self.cursor.execute(query, (TAG, ))
+                    produto = self.cursor.fetchall()
+                    print(produto)
+                    if produto:
+                        return produto
+                    else:
+                        print("Produto não encontrado.")
+                        return []
+                except:
+                    query = f"SELECT * FROM {self.INFOS['table']} WHERE NOME LIKE %s"
+                    self.cursor.execute(query, (TAG, ))
+                    produto = self.cursor.fetchall()
+                    print(produto)
+                    if produto:
+                        return produto
+                    else:
+                        print("Produto não encontrado.")
+                        return []
 
-        self.cursor.execute(f"SELECT NOME, PRECO FROM {self.INFOS["table"]}")
-        produtos = self.cursor.fetchall()
-        return produtos
+            self.cursor.execute(f"SELECT NOME, PRECO FROM {self.INFOS['table']}")
+            produtos = self.cursor.fetchall()
+            if not produtos:
+                print("Nenhum produto encontrado na base de dados.")
+            return produtos
+        except Exception as ERR:
+            print(f"Erro ao visualizar produtos: {ERR}")
+            return []
 
+    # Retorna o resultado de todos os produtos cadastrados na ordem = ID - NOME - PRECO DE VENDA
     def fazerPesquisa(self):
         self.cursor.execute(f"SELECT ID, NOME, PRECO FROM {self.INFOS["table"]}")
         result = self.cursor.fetchall()
         return result
 
+    # Verifica se um produto passado como parametro já existe no banco de dados
     def produtoJaCadastrado(self, NOME_PRODUTO_VERIFICAR:str):
         query = f"SELECT * FROM {self.INFOS["table"]} WHERE NOME LIKE %s"
         self.cursor.execute(query, (NOME_PRODUTO_VERIFICAR,))
@@ -49,9 +64,10 @@ class DATABASE():
             return True
         else:
             return False
-
+        
+    # Função para cadastrar o produto ao banco de dados
     def cadastrarProdutoDB(self, NOME_PRODUTO: str, DESCRICAO_PRODUTO: str, PRECO_VENDA_PRODUTO: float):
-
+        #Verifica se já existe o produto no banco de dados ao chamar a função produtoJaCadastrado passando produto como parametro
         if self.produtoJaCadastrado(NOME_PRODUTO_VERIFICAR=NOME_PRODUTO):
                 print(f"{NOME_PRODUTO} já está cadastrado no banco de dados!")
                 return
@@ -67,40 +83,47 @@ class DATABASE():
             
         return None
     
-    def deletarProduto(self, SEARCH_PARAMETER_TO_DELETE):
+    def deletarProduto(self, SEARCH_PARAMETER_TO_DELETE) -> bool:
         SEARCH_PARAMETER_TO_DELETE.strip()
+        if not self.visualizarProdutos(SEARCH_PARAMETER_TO_DELETE):
+            print(f"Produto '{SEARCH_PARAMETER_TO_DELETE}' não encontrado.")
+            return False
         try:
             ID_PRODUTO = int(SEARCH_PARAMETER_TO_DELETE)
             QUERY = F"DELETE FROM {self.INFOS['table']} WHERE ID = %s"
             VALUES = (ID_PRODUTO,)
-            self.cursor.execute(QUERY, VALUES)
         except ValueError:
-            QUERY = F"DELETE FROM %s WHERE NOME LIKE %s"
-            VALUES = (self.INFOS["table"], SEARCH_PARAMETER_TO_DELETE)
-            self.cursor.execute(QUERY, VALUES)
+            QUERY = F"DELETE FROM {self.INFOS["table"]} WHERE NOME LIKE %s"
+            VALUES = (SEARCH_PARAMETER_TO_DELETE, )
 
+        self.cursor.execute(QUERY, VALUES)
         self.DB.commit()
         print(F"PRODUTO: {SEARCH_PARAMETER_TO_DELETE} excluído com sucesso.")
+        return True
 
+    # Função para atualizar valores de um produto cadastrado
     def atualizarProduto(self, PRODUTO, PRODUTO_ATUALIZADO):
-        NOME_ANTIGO=PRODUTO[0][1]
-        DESCRICAO_ANTIGA=PRODUTO[0][2]
-        VALOR_ANTIGO=PRODUTO[0][3]
+        if len(PRODUTO) == 0:
+            print("Produto não encontrado!")
+            return False
+            
+        NOME_ANTIGO = PRODUTO[0][1]
+        DESCRICAO_ANTIGA = PRODUTO[0][2]
+        VALOR_ANTIGO = PRODUTO[0][3]
 
         novo_nome = PRODUTO_ATUALIZADO["NOME"]
         novo_valor = PRODUTO_ATUALIZADO["VALOR"]
         nova_descricao = PRODUTO_ATUALIZADO["DESCRICAO"]
 
-        query = "UPDATE PRODUTO SET NOME = %s, DESCRICAO = %s, PRECO = %s WHERE NOME = %s AND DESCRICAO = %s AND PRECO = %s"
-        valores = (novo_nome, nova_descricao, novo_valor, NOME_ANTIGO, DESCRICAO_ANTIGA, VALOR_ANTIGO)
-
-        self.cursor.execute(query, valores)
-
-        self.DB.commit()
-        print("Produto atualizado com sucesso!")
-    
+        try:
+            query = "UPDATE PRODUTO SET NOME = %s, DESCRICAO = %s, PRECO = %s WHERE NOME = %s AND DESCRICAO = %s AND PRECO = %s"
+            valores = (novo_nome, nova_descricao, novo_valor, NOME_ANTIGO, DESCRICAO_ANTIGA, VALOR_ANTIGO)
+            self.cursor.execute(query, valores)
+            self.DB.commit()
+        except Exception as e:
+            raise ValueError("Os valores inseridos são inválidos.")
         
-db = DATABASE(PASSWD="@Ninazoemayla32290282", DATABASE="PRODUTOS", TABLE="PRODUTO")
-db.produtoJaCadastrado(NOME_PRODUTO_VERIFICAR="teste")
+        return True
 
+db = DATABASE(PASSWD="@Ninazoemayla32290282", DATABASE="PRODUTOS", TABLE="PRODUTO")
